feat: add lade support (#173)
