+++ meta-cli/src/cli/serialize.rs
@@ -6,30 +6,12 @@ use crate::typegraph::TypegraphLoader;
 use crate::utils::ensure_venv;
 use anyhow::bail;
 use anyhow::{Context, Result};
+use async_trait::async_trait;
 use clap::Parser;
 use core::fmt::Debug;
-use std::cell::Cell;
-use std::fs::File;
 use std::io::{self, Write};
-use std::ops::Deref;
 use std::path::{Path, PathBuf};
-
-#[derive(Default)]
-struct Writer(Cell<Option<Box<dyn Write>>>);
-
-impl Debug for Writer {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "Writer")
-    }
-}
-
-impl Deref for Writer {
-    type Target = Cell<Option<Box<dyn Write>>>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
+use tokio::io::AsyncWriteExt;
 
 #[derive(Parser, Debug)]
 pub struct Serialize {
@@ -50,15 +32,13 @@ pub struct Serialize {
     #[clap(short, long, value_parser)]
     out: Option<String>,
 
-    #[clap(skip)]
-    writer: Writer,
-
     #[clap(long, default_value_t = false)]
     pretty: bool,
 }
 
+#[async_trait]
 impl Action for Serialize {
-    fn run(&self, dir: String, config_path: Option<PathBuf>) -> Result<()> {
+    async fn run(&self, dir: String, config_path: Option<PathBuf>) -> Result<()> {
         ensure_venv(&dir)?;
 
         // config file is not used when `TypeGraph` files
@@ -91,46 +71,41 @@ impl Action for Serialize {
 
         if let Some(tg_name) = self.typegraph.as_ref() {
             if let Some(tg) = tgs.into_iter().find(|tg| &tg.name().unwrap() == tg_name) {
-                self.write(&self.to_string(&tg)?)?;
+                self.write(&self.to_string(&tg)?).await?;
             } else {
                 bail!("typegraph \"{}\" not found", tg_name);
             }
         } else if self.unique {
             if tgs.len() == 1 {
-                self.write(&self.to_string(&tgs[0])?)?;
+                self.write(&self.to_string(&tgs[0])?).await?;
             } else {
                 eprintln!("expected only one typegraph, got {}", tgs.len());
                 std::process::exit(1);
             }
         } else {
-            self.write(&self.to_string(&tgs)?)?;
+            self.write(&self.to_string(&tgs)?).await?;
         }
 
-        self.write("\n")?;
         Ok(())
     }
 }
 
 impl Serialize {
-    fn write(&self, contents: &str) -> Result<()> {
-        if let Some(mut writer) = self.writer.take() {
-            writer.write_all(contents.as_bytes())?;
-            self.writer.set(Some(writer));
-            Ok(())
+    async fn write(&self, contents: &str) -> Result<()> {
+        if let Some(path) = self.out.as_ref() {
+            tokio::fs::OpenOptions::new()
+                .truncate(true)
+                .create(true)
+                .write(true)
+                .open(path)
+                .await
+                .context("Could not open file")?
+                .write_all(contents.as_bytes())
+                .await?;
         } else {
-            if let Some(path) = self.out.as_ref() {
-                let file = File::options()
-                    .truncate(true)
-                    .create(true)
-                    .write(true)
-                    .open(path)
-                    .context("Could not open file")?;
-                self.writer.set(Some(Box::new(file)));
-            } else {
-                self.writer.set(Some(Box::new(io::stdout())));
-            }
-            self.write(contents)
-        }
+            io::stdout().write_all(contents.as_bytes())?;
+        };
+        Ok(())
     }
 
     fn to_string<T: serde::Serialize>(&self, val: &T) -> serde_json::Result<String> {
