+++ typegate/src/typegraphs/prisma_migration.json
@@ -109,7 +109,7 @@
     },
     {
       "type": "function",
-      "title": "function_23",
+      "title": "function_24",
       "runtime": 1,
       "policies": [
         0
@@ -123,7 +123,7 @@
     },
     {
       "type": "object",
-      "title": "object_7",
+      "title": "object_19",
       "runtime": 1,
       "policies": [],
       "config": {},
@@ -183,7 +183,7 @@
     },
     {
       "type": "object",
-      "title": "object_22",
+      "title": "object_23",
       "runtime": 1,
       "policies": [],
       "config": {},
@@ -195,14 +195,14 @@
     },
     {
       "type": "boolean",
-      "title": "boolean_19",
+      "title": "boolean_20",
       "runtime": 1,
       "policies": [],
       "config": {}
     },
     {
       "type": "array",
-      "title": "array_21",
+      "title": "array_22",
       "runtime": 1,
       "policies": [],
       "config": {},
@@ -210,14 +210,14 @@
     },
     {
       "type": "string",
-      "title": "string_20",
+      "title": "string_21",
       "runtime": 1,
       "policies": [],
       "config": {}
     },
     {
       "type": "function",
-      "title": "function_32",
+      "title": "function_34",
       "runtime": 1,
       "policies": [
         0
@@ -231,7 +231,7 @@
     },
     {
       "type": "object",
-      "title": "object_7",
+      "title": "object_27",
       "runtime": 1,
       "policies": [],
       "config": {},
@@ -246,21 +246,21 @@
     },
     {
       "type": "string",
-      "title": "string_24",
+      "title": "string_25",
       "runtime": 1,
       "policies": [],
       "config": {}
     },
     {
       "type": "boolean",
-      "title": "boolean_25",
+      "title": "boolean_26",
       "runtime": 1,
       "policies": [],
       "config": {}
     },
     {
       "type": "object",
-      "title": "object_31",
+      "title": "object_33",
       "runtime": 1,
       "policies": [],
       "config": {},
@@ -274,25 +274,18 @@
     },
     {
       "type": "string",
-      "title": "string_26",
+      "title": "string_28",
       "runtime": 1,
       "policies": [],
       "config": {}
     },
     {
       "type": "array",
-      "title": "array_28",
+      "title": "array_30",
       "runtime": 1,
       "policies": [],
       "config": {},
       "items": 30
-    },
-    {
-      "type": "string",
-      "title": "string_27",
-      "runtime": 1,
-      "policies": [],
-      "config": {}
     },
     {
       "type": "string",
@@ -303,14 +296,21 @@
     },
     {
       "type": "string",
-      "title": "string_30",
+      "title": "string_31",
+      "runtime": 1,
+      "policies": [],
+      "config": {}
+    },
+    {
+      "type": "string",
+      "title": "string_32",
       "runtime": 1,
       "policies": [],
       "config": {}
     },
     {
       "type": "function",
-      "title": "function_38",
+      "title": "function_41",
       "runtime": 1,
       "policies": [
         0
@@ -324,7 +324,7 @@
     },
     {
       "type": "object",
-      "title": "object_7",
+      "title": "object_36",
       "runtime": 1,
       "policies": [],
       "config": {},
@@ -337,14 +337,14 @@
     },
     {
       "type": "string",
-      "title": "string_33",
+      "title": "string_35",
       "runtime": 1,
       "policies": [],
       "config": {}
     },
     {
       "type": "object",
-      "title": "object_37",
+      "title": "object_40",
       "runtime": 1,
       "policies": [],
       "config": {},
@@ -356,14 +356,14 @@
     },
     {
       "type": "integer",
-      "title": "integer_34",
+      "title": "integer_37",
       "runtime": 1,
       "policies": [],
       "config": {}
     },
     {
       "type": "array",
-      "title": "array_36",
+      "title": "array_39",
       "runtime": 1,
       "policies": [],
       "config": {},
@@ -371,7 +371,7 @@
     },
     {
       "type": "string",
-      "title": "string_35",
+      "title": "string_38",
       "runtime": 1,
       "policies": [],
       "config": {}
