+++ typegate/src/typegraphs/typegate.json
@@ -75,7 +75,7 @@
     },
     {
       "type": "function",
-      "title": "function_14",
+      "title": "function_15",
       "runtime": 1,
       "policies": [
         0
@@ -107,7 +107,7 @@
     },
     {
       "type": "optional",
-      "title": "optional_13",
+      "title": "optional_14",
       "runtime": 1,
       "policies": [],
       "config": {},
@@ -116,7 +116,7 @@
     },
     {
       "type": "object",
-      "title": "typegraph",
+      "title": "object_13",
       "runtime": 1,
       "policies": [],
       "config": {},
@@ -157,7 +157,7 @@
     },
     {
       "type": "function",
-      "title": "function_18",
+      "title": "function_19",
       "runtime": 1,
       "policies": [
         0
@@ -171,7 +171,7 @@
     },
     {
       "type": "object",
-      "title": "object_16",
+      "title": "object_17",
       "runtime": 1,
       "policies": [],
       "config": {},
@@ -182,14 +182,14 @@
     },
     {
       "type": "string",
-      "title": "string_15",
+      "title": "string_16",
       "runtime": 1,
       "policies": [],
       "config": {}
     },
     {
       "type": "optional",
-      "title": "optional_17",
+      "title": "optional_18",
       "runtime": 1,
       "policies": [],
       "config": {},
@@ -198,7 +198,7 @@
     },
     {
       "type": "function",
-      "title": "function_22",
+      "title": "function_23",
       "runtime": 1,
       "policies": [
         0
@@ -212,7 +212,7 @@
     },
     {
       "type": "object",
-      "title": "object_20",
+      "title": "object_21",
       "runtime": 1,
       "policies": [],
       "config": {},
@@ -223,14 +223,14 @@
     },
     {
       "type": "string",
-      "title": "string_19",
+      "title": "string_20",
       "runtime": 1,
       "policies": [],
       "config": {}
     },
     {
       "type": "integer",
-      "title": "integer_21",
+      "title": "integer_22",
       "runtime": 1,
       "policies": [],
       "config": {}
