+++ website/docs/reference/typegate/prisma-migration/objects.md
@@ -43,9 +43,9 @@ diff<br />
 </tbody>
 </table>
 
-## object_22
+## object_23
 
-object_22 type
+object_23 type
 
 <p style={{ marginBottom: "0.4em" }}><strong>Fields</strong></p>
 
@@ -81,9 +81,9 @@ databaseReset<br />
 </tbody>
 </table>
 
-## object_31
+## object_33
 
-object_31 type
+object_33 type
 
 <p style={{ marginBottom: "0.4em" }}><strong>Fields</strong></p>
 
@@ -145,9 +145,9 @@ createdMigrationName<br />
 </tbody>
 </table>
 
-## object_37
+## object_40
 
-object_37 type
+object_40 type
 
 <p style={{ marginBottom: "0.4em" }}><strong>Fields</strong></p>
