+++ website/docs/reference/typegate/typegate/objects.md
@@ -5,6 +5,57 @@ slug: objects
 sidebar_position: 4
 ---
 
+## object_13
+
+object_13 type
+
+<p style={{ marginBottom: "0.4em" }}><strong>Fields</strong></p>
+
+<table>
+<thead><tr><th>Name</th><th>Description</th></tr></thead>
+<tbody>
+<tr>
+<td>
+serialized<br />
+<a href="/docs/reference/typegate/typegate/scalars#string"><code>String!</code></a>
+</td>
+<td>
+<p>serialized field</p>
+<p>Policies:</p>
+<ul>
+<li>inherit</li>
+</ul>
+</td>
+</tr>
+<tr>
+<td>
+url<br />
+<a href="/docs/reference/typegate/typegate/scalars#string"><code>String!</code></a>
+</td>
+<td>
+<p>url field</p>
+<p>Policies:</p>
+<ul>
+<li>inherit</li>
+</ul>
+</td>
+</tr>
+<tr>
+<td>
+name<br />
+<a href="/docs/reference/typegate/typegate/scalars#string"><code>String!</code></a>
+</td>
+<td>
+<p>name field</p>
+<p>Policies:</p>
+<ul>
+<li>inherit</li>
+</ul>
+</td>
+</tr>
+</tbody>
+</table>
+
 ## typegraph
 
 typegraph type
