+++ website/vale/Vocab/Metatype/accept.txt
@@ -32,3 +32,5 @@ CORS
 OAuth2
 Deno
 sandboxed
+Infisical
+Zsh
